+++ lib/src/us/kbase/kbobjectutilities/KbObjectUtilitiesClient.java
@@ -284,6 +284,23 @@ public class KbObjectUtilitiesClient {
     }
 
     /**
+     * <p>Original spec-file function name: KButil_update_genome_lineage_from_files</p>
+     * <pre>
+     * </pre>
+     * @param   params   instance of type {@link us.kbase.kbobjectutilities.KButilUpdateGenomeLineageFromFilesParams KButilUpdateGenomeLineageFromFilesParams} (original type "KButil_update_genome_lineage_from_files_Params")
+     * @return   instance of type {@link us.kbase.kbobjectutilities.KButilUpdateGenomeLineageFromFilesOutput KButilUpdateGenomeLineageFromFilesOutput} (original type "KButil_update_genome_lineage_from_files_Output")
+     * @throws IOException if an IO exception occurs
+     * @throws JsonClientException if a JSON RPC exception occurs
+     */
+    public KButilUpdateGenomeLineageFromFilesOutput kButilUpdateGenomeLineageFromFiles(KButilUpdateGenomeLineageFromFilesParams params, RpcContext... jsonRpcContext) throws IOException, JsonClientException {
+        List<Object> args = new ArrayList<Object>();
+        args.add(params);
+        TypeReference<List<KButilUpdateGenomeLineageFromFilesOutput>> retType = new TypeReference<List<KButilUpdateGenomeLineageFromFilesOutput>>() {};
+        List<KButilUpdateGenomeLineageFromFilesOutput> res = caller.jsonrpcCall("kb_ObjectUtilities.KButil_update_genome_lineage_from_files", args, retType, true, true, jsonRpcContext, this.serviceVersion);
+        return res.get(0);
+    }
+
+    /**
      * <p>Original spec-file function name: KButil_update_genome_features_from_file</p>
      * <pre>
      * </pre>